Andrew/graphics refactor (#77)
* Graphics Refactor Finished

Moved functionality to different files to start graphics refactor process.

* Fixed stuff from master on this commit

* Crunch changes

* Andrew: Fixed an issue where IsDeviceSuitable wasn't being called on all surfaces.

* Fixed warnings

* Removed the old files

* Fixes requested on pull request for generics folder

* Fixes for vulkan folder

* Added changes requested for events in Graphics System

* Fixed the camera not moving in the LambPlanet

* Release warnings fixes, still an issue with the editor

* Kiss_fft fix

* Fixed more requests on PR

* Did the reorder of the events for drawing

+++ Dependencies/crunch/inc/crn_decomp.h
@@ -13,6 +13,8 @@
 #ifndef CRND_INCLUDE_CRND_H
 #define CRND_INCLUDE_CRND_H
 
+#include <cstdint>
+
 // Include crnlib.h (only to bring in some basic CRN-related types).
 #include "crnlib.h"
 
@@ -2530,7 +2532,7 @@ namespace crnd
          return NULL;
       }
 
-      CRND_ASSERT(((uint32)p_new & (CRND_MIN_ALLOC_ALIGNMENT - 1)) == 0);
+      CRND_ASSERT(((uintptr_t)p_new & (CRND_MIN_ALLOC_ALIGNMENT - 1)) == 0);
 
       return p_new;
    }
@@ -2555,7 +2557,7 @@ namespace crnd
       if (pActual_size)
          *pActual_size = actual_size;
 
-      CRND_ASSERT(((uint32)p_new & (CRND_MIN_ALLOC_ALIGNMENT - 1)) == 0);
+      CRND_ASSERT(((uintptr_t)p_new & (CRND_MIN_ALLOC_ALIGNMENT - 1)) == 0);
 
       return p_new;
    }
